% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adinar.R
\name{adinar}
\alias{adinar}
\title{AdINAR model training}
\usage{
adinar(
  time_series,
  prior = list(a_alpha = NULL, b_alpha = NULL, a_lambda = NULL, b_lambda = NULL,
    a_theta = NULL, b_theta = NULL, a_w = NULL, b_w = NULL),
  burn_in = 10^3,
  chain_length = 10^4,
  random_seed = 1761,
  verbose = TRUE
)
}
\arguments{
\item{time_series}{A univariate time series.}

\item{prior}{List of prior hyperparameters:
\describe{
  \item{a_alpha, b_alpha}{Hyperparameters of the thinning component.}
  \item{a_lambda, b_lambda}{Hyperparameters of the Gamma prior for the Poisson innovation rate.}
  \item{a_theta, b_theta}{Hyperparameters of the Beta prior for the Geometric parameter.}
  \item{a_w, b_w}{Hyperparameters of the Beta prior for the Geometric-Poisson weight mixture.}
}}

\item{burn_in}{Length of the "burn-in" period.}

\item{chain_length}{Length of the Markov chain.}

\item{random_seed}{Value of the random seed.}

\item{verbose}{Displays running status if \code{TRUE}.}
}
\value{
\code{adinar} returns an object of class "adinar".
}
\description{
Computes the posterior distribution for the AdINAR model using a Gibbs sampler.
}
