% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinar.R
\name{dpinar}
\alias{dpinar}
\title{DP-INAR model training}
\usage{
dpinar(
  time_series,
  prior = list(a_alpha = NULL, b_alpha = NULL, a_tau = NULL, b_tau = NULL, a0 = NULL,
    b0 = NULL, lambda_max = NULL),
  burn_in = 10^3,
  chain_length = 10^4,
  random_seed = 1761,
  verbose = TRUE
)
}
\arguments{
\item{time_series}{A univariate time series.}

\item{prior}{List of prior hyperparameters:
\describe{
  \item{a_alpha, b_alpha}{Hyperparameters of the thinning component.}
  \item{a_tau, b_tau}{Hyperparameters of the concentration parameter with Gamma prior.}
  \item{a0, b0}{Base measure hyperparameters.}
  \item{lambda_max}{Hyperparameter of the uniform distribution that minimizes the corresponding D-KL.}
}}

\item{burn_in}{Length of the "burn-in" period.}

\item{chain_length}{Length of the Markov chain.}

\item{random_seed}{Value of the random seed.}

\item{verbose}{Displays running status if \code{TRUE}.}
}
\value{
\code{dpinar} returns an object of class "dpinar".
}
\description{
Computes the posterior distribution for the DP-INAR family using a Gibbs sampler.
}
