% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{inar}
\alias{inar}
\title{INAR model training}
\usage{
inar(
  time_series,
  prior = list(a_alpha = NULL, b_alpha = NULL, a_lambda = NULL, b_lambda = NULL),
  burn_in = 10^3,
  chain_length = 10^4,
  random_seed = 1761,
  verbose = TRUE
)
}
\arguments{
\item{time_series}{A univariate time series.}

\item{prior}{List of prior hyperparameters:
\describe{
  \item{a_alpha, b_alpha}{Hyperparameters of the thinning component.}
  \item{a_lambda, b_lambda}{Hyperparameters of the Gamma prior for the Poisson innovation rate.}
}}

\item{burn_in}{Length of the "burn-in" period.}

\item{chain_length}{Length of the Markov chain.}

\item{random_seed}{Value of the random seed.}

\item{verbose}{Displays running status if \code{TRUE}.}
}
\value{
\code{inar} returns an object of class "inar".
}
\description{
Computes the posterior distribution for the INAR model using a Gibbs sampler.
}
