% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pittsburgh.R
\docType{data}
\name{Pittsburgh}
\alias{Pittsburgh}
\title{Pittsburgh burglary data set.}
\format{
A data frame with 144 months and 36 patrol areas.
}
\source{
\url{http://www.forecastingprinciples.com/index.php/crimedata}
}
\usage{
Pittsburgh
}
\description{
A dataset containing the reported burglary cases in Pitttsburgh, USA, from January 1990 to December 2001.
}
\keyword{datasets}
