% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adinar.R
\name{predict.adinar}
\alias{predict.adinar}
\title{Predict method for AdINAR models}
\usage{
\method{predict}{adinar}(model, h = 1, replications = 10^3)
}
\arguments{
\item{model}{An object of class "adinar".}

\item{h}{Number of steps ahead to be predicted.}

\item{replications}{Number of replications for each posterior sample.}
}
\value{
A list with the following elements: 
\describe{
  \item{est}{The \code{h}-steps-ahead prediction.}
  \item{distr}{The \code{h}-steps-ahead predictive distribution.}
}
}
\description{
Predictions and predictive distribution from a trained AdINAR model object.
}
