% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyinar.R
\name{pyinar}
\alias{pyinar}
\title{Model training}
\usage{
pyinar(time_series, p = 1, prior = list(a_alpha = NULL, a0 = NULL, b0 =
  NULL, tau = NULL, k0 = NULL, sigma = NULL, lambda_max = NULL),
  burn_in = 10^3, chain_length = 10^4, random_seed = 1761,
  verbose = TRUE)
}
\value{
A model
}
\description{
Computes the posterior distribution
}
