% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinar.R
\name{summary.dpinar}
\alias{summary.dpinar}
\title{DP-INAR model summaries
Summarizes a trained DP-INAR model}
\usage{
\method{summary}{dpinar}(model)
}
\value{
A model summary.
}
\description{
DP-INAR model summaries
Summarizes a trained DP-INAR model
}
