% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{summary.inar}
\alias{summary.inar}
\title{INAR model summaries
Summarizes a trained INAR model.}
\usage{
\method{summary}{inar}(model)
}
\value{
A model summary.
}
\description{
INAR model summaries
Summarizes a trained INAR model.
}
