\name{BayesINAR-package}
\alias{BayesINAR-package}
\alias{BayesINAR}
\docType{package}
\title{
  Bayesian INAR models
}
\description{
  Generalized Bayesian models for time series of counts, with functions for
  training and prediction.
}
% \details{
%   This section should provide a more detailed overview of how to use the
%   package, including the most important functions.
% }
\author{
  Helton Graziadei, Paulo C. Marques F. and Hedibert F. Lopes

  Maintainer: Paulo C. Marques F. <PauloCMF1@insper.edu.br>
}
\references{
  Paulo C. Marques F., Helton Graziadei and Hedibert Lopes.
  "Bayesian generalizations of the integer-valued autoregressive model."
  Journal of Applied Statistics, v. 1, p. 1-21, 2020.
}
\keyword{package}
% \seealso{
%   Optional links to other man pages
% }
% \examples{
%   \dontrun{
%      ## Optional simple examples of the most important functions
%      ## These can be in \dontrun{} and \donttest{} blocks.   
%   }
% }
