% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Time series cross-validation with generalized Bayesian INAR models}
\usage{
cross_validate(time_series, model = "inar", p = 1, h = 1,
  training_epoch = NULL, prior = list(a_alpha = NULL, a0 = NULL, b0 =
  NULL, tau = NULL, k0 = NULL, a_tau = NULL, b_tau = NULL, a_w = NULL, b_w
  = NULL, a_theta = NULL, b_theta = NULL, sigma = NULL, lambda_max = NULL),
  burn_in = 10^3, chain_length = 10^4, random_seed = 1761,
  verbose = TRUE)
}
\arguments{
\item{time_series}{A univariate time series.}

\item{h}{Number of steps ahead to be predicted.}

\item{training_epoch}{The last observation of the first training set.}

\item{prior}{List of prior hyperparameters where:
\describe{
 \item{a_alpha}{Hyperparameters of the thinning component.}
 \item{a_theta, b_theta}{Hyperparameters of the Geometric innovation parameter.}
 \item{a_lambda, b_lambda}{Hyperparameters of the Poisson innovation rate.}
 \item{a_w, b_w}{Hyperparameters of the Geometric-Poisson mixture weight.}
 \item{a_tau, b_tau}{Hyperparameters of the concentration parameter Gamma prior.}
 \item{a0, b0}{Base measure hyperparameters.}
 \item{lambda_max}{Hyperparameter of the uniform distribution that minimizes the corresponding D-KL.}
}}

\item{burn_in}{Number of iterations for the "burn-in" period which are discarded in the chain.}

\item{chain_length}{Number of iterations of the chain.}

\item{random_seed}{Value of the random seed generator.}

\item{verbose}{If \code{TRUE} log info is provided.}
}
\value{
A list with the following elements:
\describe{
 \item{est}{Predictions of the test set.}
 \item{mae}{Mean Absolute Error of the test set predictions.}
}
}
\description{
Obtain predictions of a given test set and the mean absolute errors
}
