% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinar.R
\name{dpinar}
\alias{dpinar}
\title{DP-INAR model training}
\usage{
dpinar(time_series, p = 1, prior = list(a_alpha = NULL, a_tau = NULL,
  b_tau = NULL, a0 = NULL, b0 = NULL, lambda_max = NULL), burn_in = 10^3,
  chain_length = 10^4, random_seed = 1761, verbose = TRUE)
}
\arguments{
\item{time_series}{A univariate time series.}

\item{prior}{List of prior hyperparameters where:
\describe{
 \item{a_alpha}{Hyperparameters of the thinning component.}
 \item{a_tau, b_tau}{Hyperparameters of the concentration parameter with Gamma prior.}
 \item{a0, b0}{Base measure hyperparameters.}
 \item{lambda_max}{Hyperparameter of the uniform distribution that minimizes the corresponding D-KL.}
}}

\item{burn_in}{Number of iterations for the "burn-in" period which are discarded in the chain.}

\item{chain_length}{Number of iterations of the chain.}

\item{random_seed}{Value of the random seed generator.}

\item{verbose}{If \code{TRUE} log info is provided.}
}
\value{
dpinar returns an object of class "dpinar".
}
\description{
Computes the posterior distribution for the DP-INAR family using a Gibbs sampler.
}
