% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{inar}
\alias{inar}
\title{Model training}
\usage{
inar(time_series, p = 1, prior = list(a_alpha = NULL, a_lambda = NULL,
  b_lambda = NULL), burn_in = 10^3, chain_length = 10^4,
  random_seed = 1761, verbose = TRUE)
}
\value{
A model
}
\description{
Computes the posterior distribution
}
