% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{predict.inar}
\alias{predict.inar}
\title{Predict Method for INAR models}
\usage{
\method{predict}{inar}(model, h = 1, replications = 10^4)
}
\arguments{
\item{model}{A trained object of class inheriting from "inar".}

\item{h}{Number of steps ahead to be predicted.}

\item{replications}{Number of replications for each posterior sample.}
}
\value{
A list with the following elements: 
\describe{
 \item{est}{The \code{h}-steps-ahead prediction.}
 \item{distr}{The \code{h}-steps-ahead predictive distribution.}
}
}
\description{
Obtains predictions and predictive distribution from a trained INAR model object.
}
